COMMERCE-8197 Renamed and moved dataset tables
